OPS-133 fix: trend 수정

+++ src/main/java/com/sprint/example/sb01part2hrbankteam10/dto/EmployeeTrendDto.java
@@ -10,8 +10,15 @@ import lombok.NoArgsConstructor;
 public class EmployeeTrendDto {
 
   private String date;
-  private Integer count;
-  private Integer change;
+  private Long count;
+  private Long change;
   private Double changeRate;
+
+//  public EmployeeTrendDto(String date, Long count, Long change, Double changeRate) {
+//    this.date = date;
+//    this.count = count;
+//    this.change = (change != null) ? change : 0;
+//    this.changeRate = (changeRate != null) ? Math.round(changeRate * 100.0) / 100.0 : 0.0;
+//  }
 
 }
